+++ app/src/main/java/com/android/finalproject/data/database/MovieDao.kt
@@ -12,5 +12,5 @@ interface MovieDao {
     fun addMovieToFav(movie: Movie): Long
 
     @Delete
-    fun removeMovieFromFav(movie: Movie): Long
+    fun removeMovieFromFav(movie: Movie): Int
 }
